backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
FastAPI-SQLAlchemy==0.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-core==2.18.1
PyJWT==2.6.0
python-decouple==3.7
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing-extensions==4.11.0
uvicorn==0.29.0
